Fix WorkspaceWelcome page: separate to content and page; and make some style
